Added the asset model.

+++ database/migrations/2018_06_02_171801_create_role_user_table.php
@@ -12,12 +12,12 @@ class CreateRoleUserTable extends Migration
     public function up()
     {
         Schema::create('role_user', function (Blueprint $table) {
-            $table->unsignedBigInteger('role_id');
+            $table->unsignedInteger('role_id');
             $table->foreign('role_id')
                 ->references('id')
                 ->on('user_roles')
                 ->onDelete('cascade');
-            $table->unsignedBigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
