+++ database/migrations/2018_06_03_110727_create_user_auth_tokens_table.php
@@ -12,8 +12,8 @@ class CreateUserAuthTokensTable extends Migration
     public function up()
     {
         Schema::create('user_auth_tokens', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('user_id');
+            $table->increments('id');
+            $table->unsignedInteger('user_id');
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
